Move code related to reading tsconfig.json into a function
